More cleanup and split into several files
