+++ vector/src/main/java/im/vector/app/features/spaces/people/SpacePeopleSharedAction.kt
@@ -16,7 +16,12 @@
 
 package im.vector.app.features.spaces.people
 
-import im.vector.app.core.platform.VectorSharedActionViewModel
-import javax.inject.Inject
+import im.vector.app.core.platform.VectorSharedAction
 
-class SpacePeopleSharedActionViewModel @Inject constructor() : VectorSharedActionViewModel<SpacePeopleSharedAction>()
+sealed class SpacePeopleSharedAction : VectorSharedAction {
+    object Dismiss : SpacePeopleSharedAction()
+    object ShowModalLoading : SpacePeopleSharedAction()
+    object HideModalLoading : SpacePeopleSharedAction()
+    data class NavigateToRoom(val roomId: String) : SpacePeopleSharedAction()
+    data class NavigateToInvite(val spaceId: String) : SpacePeopleSharedAction()
+}
